Fix StealHPFIx Duplication Merge Teleport commands into 1 Fix Npcs unspawn command

+++ Server/Stump.Server.WorldServer/Game/Maps/Map.cs
@@ -500,7 +500,7 @@ namespace Stump.Server.WorldServer.Game.Maps
             return npc;
         }
 
-        public bool UnSpawnNpc(sbyte id)
+        public bool UnSpawnNpc(short id)
         {
             var npc = GetActor<Npc>(id);
